segundo commit

+++ producto-service/src/main/java/com/negocio/producto_service/service/ProductoService.java
@@ -1,6 +1,5 @@
 package com.negocio.producto_service.service;
 
-import java.util.ArrayList;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -20,19 +19,25 @@ public class ProductoService implements IProductoService {
     private ICategoriaRepository repoCategoria;
 
     @Override
-    public String createProducto(Long id_producto, String nombre, String descripcion, Long id_categoria) {
-        Categoria categoria = repoCategoria.findById(id_categoria).orElse(null);
+    public String createProducto(Long id_producto, String nombre, Integer cantidad, String descripcion,
+            Long id_categoria,Double precio) {
+        Categoria categoria = repoCategoria.findByNombre(nombre);
         Producto producto = new Producto();
         producto.setId_producto(id_producto);
         producto.setNombre(nombre);
         producto.setDescripcion(descripcion);
         producto.setCategoria(categoria);
+        producto.setCantidad(cantidad);
+        producto.setPrecio(precio);
         repoProducto.save(producto);
         return "Producto guardado con exito";
     }
 
     @Override
     public String deleteProduto(Long id) {
+        if(id.equals(null)){
+            return "El producto no existe";
+        }
         repoProducto.deleteById(id);
 
         return "Producto eliminado con exito";
